styling and refactor
