+++ frontend/src/components/index.js
@@ -2,4 +2,3 @@ export { Loader } from "./Loader";
 export { Main } from "./Layout/Main";
 export { Sidebar } from "./Sidebar";
 export { Accordion } from "./Accordion";
-export { TitleCard } from "./TitleCard";
